+++ packages/design-system/infinite-scroll-list/index.ts
@@ -1,0 +1,1 @@
+export * from "./infinite-scroll-list";
