+++ packages/design-system/tooltip/tooltip.web.tsx
@@ -5,7 +5,6 @@ import * as Portal from "@radix-ui/react-portal";
 import { AnimatePresence, MotiView } from "moti";
 
 import { useIsMobileWeb } from "@showtime-xyz/universal.hooks";
-import { tw as tailwind } from "@showtime-xyz/universal.tailwind";
 
 import { Position } from "./position";
 import { TooltipContent } from "./tooltop.content";
@@ -55,7 +54,7 @@ export const Tooltip: React.FC<TooltipProps> = ({
       {/* @ts-ignore */}
       <AnimatePresence>
         {show && (
-          <Portal.Root style={tailwind.style(`w-full`)}>
+          <Portal.Root style={{ width: "100%" }}>
             <MotiView
               from={{
                 opacity: 0,
